+++ server/src/db/migrate/defaults.ts
@@ -66,6 +66,7 @@ import * as clifreeFleetBenchmarkSlug from '../migrations/20260922_000002_clifre
 import * as clifreeFleetUsage from '../migrations/20260922_000003_clifree_fleet_usage.js';
 import * as clifreeFleetUsagePerDay from '../migrations/20260923_000001_clifree_fleet_usage_per_day.js';
 import * as chainMinimumsAndAaHistory from '../migrations/20260925_000001_chain_minimums_and_aa_history.js';
+import * as quotaPolicyUnit from '../migrations/20260925_000002_quota_policy_unit.js';
 import * as requestModelAttribution from '../migrations/20260913_000001_request_model_attribution.js';
 
 export interface MigrationModule {
@@ -215,4 +216,5 @@ export const DEFAULT_MIGRATIONS: readonly DefaultMigration[] = [
   { filename: '20260922_000003_clifree_fleet_usage.ts', module: clifreeFleetUsage },
   { filename: '20260923_000001_clifree_fleet_usage_per_day.ts', module: clifreeFleetUsagePerDay },
   { filename: '20260925_000001_chain_minimums_and_aa_history.ts', module: chainMinimumsAndAaHistory },
+  { filename: '20260925_000002_quota_policy_unit.ts', module: quotaPolicyUnit },
 ];